Using a ZipStream is just not working out.  The comment field does not appear to be being written to the stream (or else the read is not picking it up).  This dog won't hunt.
Rather than keep going down this road, I'm going to switch horses and use XML to serialize the data.  At some point in the future I may incoporate zip to reduce the size of the XML but not in the first go around.

I'm checking this in so that I have a record of what I have done in case I need to get back to it.  Expect the next revision to delete a lot of this.