feature: Added support for HTML5 Submissions.
FormsFeeder now handles submissions from HTML5 Forms.  It expects that there will be a <FormFeeder> element with a <Plugin> element inside it (i.e. there will be a "FormFeeder" subform with a "Plugin" field inside).  It calls the plugin with the name provided in the Plugin field and expects the plugin to return 0 or 1 datasources.  If a datasource is returned, then it returns that datasource in the response.