Changes to support absolute paths.
HTML5 and Adaptive Forms can now contain absolute paths in the
dependency links (Link and Script Elements) instead of relative links
(which is the default).

Unfortunately, this is split into two parts:
1) the HTML5/AF APIs have a UrlFilter that is used to replace the URLs
in the form itself (see the absoluteLocation method in
FormsFeederUrlFilterBuilder class)

2) the FormFeeder AEM proxy code must replace a line in utils.js that
determines the context.  This is done using an application.property on
the server

These two setting must match and must be both turned on for absolute
paths to work.

I'd like to see this improved at some point in the future, but for now,
things seem to work.