Updated pf4j and plugins to use the latest version of Spring.
Also added git-commit-id-plugin to plugins project.  This will create a
git.properties inside the plugin's .jar file which will help in
identifying the .jar.   At some point in the future, I would like the
plugin loader to look for and print out the build id of any plugins it
loads.